+++ packages/std/src/results/cosmos_msg.rs
@@ -22,56 +22,72 @@ where
     Wasm(WasmMsg),
 }
 
+/// The message types of the bank module.
+///
+/// See https://github.com/cosmos/cosmos-sdk/blob/v0.40.0/proto/cosmos/bank/v1beta1/tx.proto
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
 pub enum BankMsg {
-    // this moves tokens in the underlying sdk
+    /// Sends native tokens from the contract to the given address.
+    ///
+    /// This is translated to a [MsgSend](https://github.com/cosmos/cosmos-sdk/blob/v0.40.0/proto/cosmos/bank/v1beta1/tx.proto#L19-L28).
+    /// `from_address` is automatically filled with the current contract's address.
     Send {
-        from_address: HumanAddr,
         to_address: HumanAddr,
         amount: Vec<Coin>,
     },
 }
 
+/// The message types of the staking module.
+///
+/// See https://github.com/cosmos/cosmos-sdk/blob/v0.40.0/proto/cosmos/staking/v1beta1/tx.proto
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
 pub enum StakingMsg {
-    Delegate {
-        // delegator is automatically set to address of the calling contract
-        validator: HumanAddr,
-        amount: Coin,
-    },
-    Undelegate {
-        // delegator is automatically set to address of the calling contract
-        validator: HumanAddr,
-        amount: Coin,
-    },
+    /// This is translated to a [MsgDelegate](https://github.com/cosmos/cosmos-sdk/blob/v0.40.0/proto/cosmos/staking/v1beta1/tx.proto#L81-L90).
+    /// `delegator_address` is automatically filled with the current contract's address.
+    Delegate { validator: HumanAddr, amount: Coin },
+    /// This is translated to a [MsgUndelegate](https://github.com/cosmos/cosmos-sdk/blob/v0.40.0/proto/cosmos/staking/v1beta1/tx.proto#L112-L121).
+    /// `delegator_address` is automatically filled with the current contract's address.
+    Undelegate { validator: HumanAddr, amount: Coin },
+    /// This is translated to a [MsgSetWithdrawAddress](https://github.com/cosmos/cosmos-sdk/blob/v0.40.0/proto/cosmos/distribution/v1beta1/tx.proto#L29-L37)
+    /// followed by a [MsgWithdrawDelegatorReward](https://github.com/cosmos/cosmos-sdk/blob/v0.40.0/proto/cosmos/distribution/v1beta1/tx.proto#L42-L50).
+    /// `delegator_address` is automatically filled with the current contract's address.
     Withdraw {
-        // delegator is automatically set to address of the calling contract
         validator: HumanAddr,
         /// this is the "withdraw address", the one that should receive the rewards
         /// if None, then use delegator address
         recipient: Option<HumanAddr>,
     },
+    /// This is translated to a [MsgBeginRedelegate](https://github.com/cosmos/cosmos-sdk/blob/v0.40.0/proto/cosmos/staking/v1beta1/tx.proto#L95-L105).
+    /// `delegator_address` is automatically filled with the current contract's address.
     Redelegate {
-        // delegator is automatically set to address of the calling contract
         src_validator: HumanAddr,
         dst_validator: HumanAddr,
         amount: Coin,
     },
 }
 
+/// The message types of the wasm module.
+///
+/// See https://github.com/CosmWasm/wasmd/blob/v0.14.0/x/wasm/internal/types/tx.proto
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
 pub enum WasmMsg {
-    /// this dispatches a call to another contract at a known address (with known ABI)
+    /// Dispatches a call to another contract at a known address (with known ABI).
+    ///
+    /// This is translated to a [MsgExecuteContract](https://github.com/CosmWasm/wasmd/blob/v0.14.0/x/wasm/internal/types/tx.proto#L68-L78).
+    /// `sender` is automatically filled with the current contract's address.
     Execute {
         contract_addr: HumanAddr,
         /// msg is the json-encoded HandleMsg struct (as raw Binary)
         msg: Binary,
         send: Vec<Coin>,
     },
-    /// this instantiates a new contracts from previously uploaded wasm code
+    /// Instantiates a new contracts from previously uploaded Wasm code.
+    ///
+    /// This is translated to a [MsgInstantiateContract](https://github.com/CosmWasm/wasmd/blob/v0.14.0/x/wasm/internal/types/tx.proto#L47-L61).
+    /// `sender` is automatically filled with the current contract's address.
     Instantiate {
         code_id: u64,
         /// msg is the json-encoded InitMsg struct (as raw Binary)
@@ -108,14 +124,9 @@ mod tests {
 
     #[test]
     fn from_bank_msg_works() {
-        let from_address = HumanAddr::from("me");
         let to_address = HumanAddr::from("you");
         let amount = coins(1015, "earth");
-        let bank = BankMsg::Send {
-            from_address,
-            to_address,
-            amount,
-        };
+        let bank = BankMsg::Send { to_address, amount };
         let msg: CosmosMsg = bank.clone().into();
         match msg {
             CosmosMsg::Bank(msg) => assert_eq!(bank, msg),